+++ src/salesforceUtil.ts
@@ -1,24 +1,39 @@
 
 import SObjectRecord from './models/sobjectRecord';
 import { stringEquals } from './util';
+import { write } from 'fs';
 
 /**
  * Create a record proxy around a SF record removing any namespace prefixes it may contain and making all fields accessible with case-insensitive keys.
  * @param record SF Like record object
  */
-export function createRecordProxy(record: SObjectRecord) : SObjectRecord {
-    const getPropertyKey = (target: SObjectRecord, name: string | number | symbol) => Object.keys(target).find(key => {
+export function createRecordProxy<T extends Object>(record: T, writable: boolean = false) : T & { [key: string] : any } {
+    const getPropertyKey = (target: T, name: string | number | symbol) => Object.keys(target).find(key => {
         return stringEquals(key, name.toString(), true) || stringEquals(removeNamespacePrefix(key), name.toString(), true);
     });
 
     return new Proxy(record, {
         get: (target, name) => {
             const key = getPropertyKey(target, name);
-            return key && target[key];
+            const data = target[key];
+            if (Array.isArray(data)) {
+                return data.map( element => createRecordProxy(element, writable) );
+            } else if (typeof data === 'object') {                
+                return createRecordProxy(data, writable);
+            }
+            return data;
+        },
+        set: (target, name, value) => {
+            if (writable) {
+                const key = getPropertyKey(target, name);
+                target[key || name] = value;
+                return true;
+            } 
+            return false;
         },
         getOwnPropertyDescriptor: (target, name) => {
             const key = getPropertyKey(target, name);
-            return key ? { configurable: true, enumerable: true, writable: false } : undefined;
+            return key ? { configurable: true, enumerable: true, writable } : undefined;
         },
         has: (target, name) => getPropertyKey(target, name) !== undefined,
         enumerate: (target) => Object.keys(target),